Improved output for SpecCSP, added createSubjFolds
SpecCSP output for ML functions is optimized if verbose = T to make
visualization of alpha easy.
Added createSubjFolds to properly handle fold splits for subjects as
caret's createFolds shows rather unexpected behavior for this scenario.
Also added factor conversion for outcome vectors so the is.numeric check
of createFolds is skipped.